Multiple Changes (See description)
Removed currently unused navmesh asset

Overhauled variables in health and weapon to store base and current values for calculating upgrades

Continued work on upgrade manager. This should be functional in it's current state but I haven't tested due to not having an interface to activate the methods. We should work on achieving this ASAP